refactor: refactor + add addresses to constant
