+++ src/enums/chainId.ts
@@ -1,0 +1,5 @@
+export enum ChainId {
+  CELO = 42220,
+  ALFAJORES = 44787,
+  BAKLAVA = 62320,
+}
