+++ src/enums/proposalState.ts
@@ -1,0 +1,10 @@
+export enum ProposalState {
+  PENDING = 0,
+  ACTIVE = 1,
+  CANCELED = 2,
+  DEFEATED = 3,
+  SUCCEEDED = 4,
+  QUEUED = 5,
+  EXPIRED = 6,
+  EXECUTED = 7,
+}
