+++ src/index.ts
@@ -3,4 +3,4 @@
 export * from './mento'
 export * from './governance'
 export * from './utils'
-export { ContractAddresses } from './types'
+export * from './constants'
