+++ src/interfaces/IChainClient.ts
@@ -1,0 +1,9 @@
+import { ethers, providers } from 'ethers'
+
+export interface IChainClient {
+  getSigner(): Promise<ethers.Signer | providers.Provider>
+  getChainId(): Promise<number>
+  populateTransaction(
+    tx: ethers.PopulatedTransaction
+  ): Promise<providers.TransactionRequest>
+}
