+++ src/interfaces/tradingLimit.ts
@@ -1,0 +1,8 @@
+export type Address = string
+
+export interface TradingLimit {
+  asset: Address
+  maxIn: number
+  maxOut: number
+  until: number
+}
