+++ src/interfaces/tradingLimitsConfig.ts
@@ -1,0 +1,8 @@
+export interface TradingLimitsConfig {
+  timestep0: number
+  timestep1: number
+  limit0: number
+  limit1: number
+  limitGlobal: number
+  flags: number
+}
