+++ src/interfaces/tradingLimitsState.ts
@@ -1,0 +1,7 @@
+export interface TradingLimitsState {
+  lastUpdated0: number
+  lastUpdated1: number
+  netflow0: number
+  netflow1: number
+  netflowGlobal: number
+}
