+++ src/types/contractAddressMap.ts
@@ -1,0 +1,5 @@
+import { ContractAddresses } from '.'
+
+export type ContractAddressMap = {
+  [chainId: number]: ContractAddresses
+}
