+++ src/types/contractAddresses.ts
@@ -1,0 +1,9 @@
+export type ContractAddresses = {
+  GovernanceFactory: string
+  Airgrab: string
+  Emission: string
+  MentoGovernor: string
+  MentoToken: string
+  TimelockController: string
+  Locking: string
+}
