+++ src/types/index.ts
@@ -1,0 +1,2 @@
+export * from './contractAddresses'
+export * from './contractAddressMap'
